+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.0.2 (--)
 
-* Fix int in table
+* Add paragraph block.
+* Fix int in table.
 
 ## 0.0.1 (2020-07-01)